evm/eof: rename eofUtils.ts to utils.ts
